backend/requirements.txt
fastapi
uvicorn
pydantic
pymongo
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
